Change rails dependency to railties
This library only depends on Rails::Railtie and Rails::Application, both
of which are defined in the railties gem. By depending on all of Rails,
this library forces applications to pull in things like ActionCable,
ActiveJob, etc. even if not used.

This library also depends on Rack, but this is a transient dependency of
Railties (railties -> actionpack -> rack).